% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_index_fits.R, R/compare_index_fits_1.R
\name{compare_index_fits}
\alias{compare_index_fits}
\title{Compare index fits across models}
\usage{
compare_index_fits(ssruns, mlabels)

compare_index_fits(ssruns, mlabels)
}
\arguments{
\item{mlabels}{}
}
\description{
Compare index fits across models

Compare cpue indices and fits to indices for multiple runs
}
