% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ages.R
\name{get_ages}
\alias{get_ages}
\title{Get Age Data and Fits from SS Model Runs}
\usage{
get_ages(ssruns, mnames)
}
\arguments{
\item{ssruns}{the Stock Synthesis runs to include in the data frame}

\item{mnames}{names of the Stock Synthesis runs}
}
\value{

}
\description{
Get Age Data and Fits from SS Model Runs
}
