% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch_data.R
\name{get_catch_data}
\alias{get_catch_data}
\title{Grab catch data for each model and make a tibble including all models}
\usage{
get_catch_data(ssruns, mlabel)
}
\arguments{
\item{mlabel}{}
}
\description{
Grab catch data for each model and make a tibble including all models
}
\examples{
get_catch_data(ssruns = ssruns,mlabel = mlabel)
}
