% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_data.R
\name{get_index_data}
\alias{get_index_data}
\title{Organize index data from multiple runs and make a tibble}
\usage{
get_index_data(ssruns, mlabels)
}
\arguments{
\item{mlabels}{}
}
\description{
Organize index data from multiple runs and make a tibble
}
