% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lengths.R
\name{get_lengths}
\alias{get_lengths}
\title{Get length data and fits from ss model runs}
\usage{
get_lengths(ssruns, mnames)
}
\arguments{
\item{ssruns}{the Stock Synthesis runs to include in the data frame}

\item{mnames}{the names of the Stock Synthesis runs}
}
\description{
Get length data and fits from ss model runs
}
