% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maturity_data.R
\name{get_maturity_data}
\alias{get_maturity_data}
\title{Pull maturity estimates from the ss models and aggregate}
\usage{
get_maturity_data(ssruns, mlabels)
}
\arguments{
\item{mlabels}{}
}
\description{
Pull maturity estimates from the ss models and aggregate
}
