% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projections.R
\name{get_projections}
\alias{get_projections}
\title{Write the data file for running projections using AFSC projection model code}
\usage{
get_projections(
  mydir,
  MyOutput,
  NagesProject,
  FirstYr,
  LastYr,
  RecruitmentAge,
  species
)
}
\arguments{
\item{species}{}
}
\description{
Write the data file for running projections using AFSC projection model code
}
