% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rec_data.R
\name{get_rec_data}
\alias{get_rec_data}
\title{Get recruitment data for each run into one tibble}
\usage{
get_rec_data(ssruns, mlabel)
}
\arguments{
\item{mlabel}{}
}
\description{
Get recruitment data for each run into one tibble
}
