% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_refpts.R
\name{get_refpts}
\alias{get_refpts}
\title{Grab the results of AFSC projections for biology and fishery ref pts}
\usage{
get_refpts(ProjDir, LastYr)
}
\arguments{
\item{LastYr}{}
}
\description{
Grab the results of AFSC projections for biology and fishery ref pts
}
