% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ssb_data.R
\name{get_ssb_data}
\alias{get_ssb_data}
\title{Get the ssb output from ss models and aggregate}
\usage{
get_ssb_data(ssruns, mlabel)
}
\arguments{
\item{ssruns}{}
}
\description{
Get the ssb output from ss models and aggregate
}
