% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_data.R
\name{get_timeseries_data}
\alias{get_timeseries_data}
\title{Grab and format any derived quantity timeseries for all models}
\usage{
get_timeseries_data(ssruns, mlabel, label_type)
}
\arguments{
\item{label_type}{}
}
\description{
Grab and format any derived quantity timeseries for all models
}
\examples{
 ssb.t<-get_timeseries_data(ssruns,mlabel,label_type = "SSB_")
 rec.t<-get_timeseries_data(ssruns,mlabel,label_type = "Recr_")
 spr.t<-get_timeseries_data(ssruns,mlabel,label_type = "SPRratio_")

}
