% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_compare.R
\name{plot_comp_compare}
\alias{plot_comp_compare}
\title{Plot comparison of fits to aggregated comp data (lengths or ages)}
\usage{
plot_comp_compare(ssruns, narea, mnames, comptype = "length")
}
\arguments{
\item{ssruns}{summary of outputs from a set of Stock Synthesis runs generated from using r4ss functions SSgetoutput followed by SSsummarize}

\item{narea}{number of separate areas for plotting length distributions (makes separate plots by area for 2 area models)}

\item{mnames}{model names for plot legend labels}

\item{comptype}{plot age or length comps? "age" "length" are valid inputs}
}
\description{
Plot comparison of fits to aggregated comp data (lengths or ages)
}
