% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comps_onearea.R
\name{plot_comps_onearea}
\alias{plot_comps_onearea}
\title{Plot comparison of fits to aggregated comp data (lengths or ages) for one area models}
\usage{
plot_comps_onearea(
  ssruns,
  narea,
  mnames,
  comptype = "length",
  saveplots = TRUE
)
}
\arguments{
\item{ssruns}{summary of outputs from a set of Stock Synthesis runs generated from using r4ss functions SSgetoutput followed by SSsummarize}

\item{narea}{number of separate areas for plotting length distributions (makes separate plots by area for 2 area models)}

\item{mnames}{model names for plot legend labels}

\item{comptype}{plot age or length comps? "age" "length" are valid inputs}

\item{saveplots}{save each plot as a .png file}
}
\description{
Plot comparison of fits to aggregated comp data (lengths or ages) for one area models
}
