% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_length_compare.R
\name{plot_length_compare}
\alias{plot_length_compare}
\title{Plot comparison of fits to aggregated length data}
\usage{
plot_length_compare(ssruns, narea, mnames)
}
\arguments{
\item{ssruns}{summary of outputs from a set of Stock Synthesis runs generated from using r4ss functions SSgetoutput followed by SSsummarize}

\item{narea}{number of separate areas for plotting length distributions (makes separate plots by area for 2 area models)}

\item{mnames}{model names for plot legend labels and}
}
\value{

}
\description{
Plot comparison of fits to aggregated length data
}
