% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_minus_spr.R
\name{plot_one_minus_spr}
\alias{plot_one_minus_spr}
\title{Plot time series of fishing intensity in terms of one minus SPR}
\usage{
plot_one_minus_spr(
  ssruns,
  mlabel,
  showlegend,
  showxlabel,
  color_option = "viridis",
  color_direction = "color_direction"
)
}
\arguments{
\item{showxlabel}{}
}
\description{
Plot time series of fishing intensity in terms of one minus SPR
}
\examples{
fishing_intensity_plot<-plot_one_minus_spr(ssruns = MasterList,mlabel = Mlabels,showxlabel = TRUE,showlegend = "yes")
}
