% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recs.R
\name{plot_recs}
\alias{plot_recs}
\title{Title}
\usage{
plot_recs(
  ssruns,
  mlabel,
  showlegend,
  showxlabel,
  color_option = viridis,
  color_direction = 1
)
}
\arguments{
\item{mlabel}{}
}
\description{
Title
}
