% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ssbs.R
\name{plot_ssbs}
\alias{plot_ssbs}
\title{Plot ssb estimates and uncertainty across model runs}
\usage{
plot_ssbs(
  ssruns,
  mlabel,
  showlegend,
  showxlabel,
  color_option = "viridis",
  color_direction = 1
)
}
\arguments{
\item{mlabel}{}
}
\description{
Plot ssb estimates and uncertainty across model runs
}
