% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time-series results from multiple Stock Synthesis 3 assessment models with uncertainty}
\usage{
plot_timeseries(
  data.t,
  xlabel,
  ylabel,
  showlegend,
  color_option = "viridis",
  color_direction = 1
)
}
\arguments{
\item{showlegend}{}
}
\description{
Plot time-series results from multiple Stock Synthesis 3 assessment models with uncertainty
}
\examples{
plot_timeseries(data.t = rec.t,xlabel= "Year",ylabel = "Age 0 (thousands)",showlegend = "none")

}
