% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_catch.R
\name{plot_total_catch}
\alias{plot_total_catch}
\title{Plot total catches. Easily extendable to plotting by areas included in a model.}
\usage{
plot_total_catch(ssruns, mlabel, showlegend, showxlabel)
}
\arguments{
\item{showxlabel}{}
}
\description{
Plot total catches. Easily extendable to plotting by areas included in a model.
}
