% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_proj_ss_1Area.R
\name{write_proj_ss_1Area}
\alias{write_proj_ss_1Area}
\title{Write projection data file for one area models}
\usage{
write_proj_ss_1Area(
  data_file = "Model1_Proj.dat",
  data = Models[[1]],
  NAGES = 30,
  FY = 1978,
  LY = 2013,
  RecAge = 1,
  species
)
}
\arguments{
\item{species}{}
}
\description{
Write projection data file for one area models
}
