% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_proj_ss_2Area.R
\name{write_proj_ss_2Area}
\alias{write_proj_ss_2Area}
\title{Write projection data file for 2 area models each with a growth morph using AFSC projection modeling code}
\usage{
write_proj_ss_2Area(
  data_file = "ForProjections.dat",
  data = replist,
  GrowthMorph = 1,
  area = 1,
  Narea = 2,
  Nmorph = 2,
  NAGES = 30,
  FY = 1978,
  LY = 2017,
  RecAge = 1,
  species
)
}
\arguments{
\item{species}{}
}
\description{
Write projection data file for 2 area models each with a growth morph using AFSC projection modeling code
}
