% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_index_fits.R
\name{compare_index_fits}
\alias{compare_index_fits}
\title{Compare index fits across models}
\usage{
compare_index_fits(ssruns, mlabels)
}
\arguments{
\item{ssruns}{a list from r4ss with output from one or more runs}

\item{mlabels}{labels for each of the runs included in ssruns list}
}
\description{
Compare index fits across models
}
\examples{
compare_index_fits(ssruns = ssgetoutputlist,mlabels = c("Model 25.1", "Model 25.2"))
}
