% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time-series results from multiple Stock Synthesis 3 assessment models with uncertainty}
\usage{
plot_timeseries(
  data.t,
  xlabel,
  ylabel,
  showlegend,
  color_option = "viridis",
  color_direction = 1
)
}
\arguments{
\item{data.t}{time series data as a data frame or tibble with Year and Value columns}

\item{xlabel}{the label for the x axis}

\item{ylabel}{the label for the y axis}

\item{showlegend}{true or false}

\item{color_option}{used within scale_color_viridis_d for changing the color theme}

\item{color_direction}{switch when needed to show the data better}
}
\description{
Plot time-series results from multiple Stock Synthesis 3 assessment models with uncertainty
}
\examples{
plot_timeseries(data.t = rec.t,xlabel= "Year",ylabel = "Age 0 (thousands)",showlegend = "none")

}
